chore(entities): wasm no longer exports strings metrics and unit tests on the rust side
