+++ entities/package.json
@@ -8,7 +8,7 @@
     "wasm:node": "wasm-pack build --target nodejs --release",
     "dev": "ts-node -T ./src/index.ts",
     "start": "node ./dist/index.js",
-    "test": "jest --roots ./dist"
+    "test": "jest --roots ./dist && cargo test"
   },
   "dependencies": {
     "lodash": "^4.17.19"
@@ -18,6 +18,7 @@
     "typescript": "^5.0.4",
     "@types/node": "^20.2.1",
     "@types/jest": "^26.0.14",
+    "@types/lodash": "^4.14.195",
     "jest": "^24.0.0"
   },
   "keywords": [],
