+++ entities/src/list-engine/wasm.ts
@@ -1,18 +1,6 @@
 import * as pkg from '../../pkg'
 import { ListEntityExtraction, ListEntityModel } from './typings'
 
-export const levenshteinSimilarity = (a: string, b: string): number => {
-  return pkg.levenshtein_sim(a, b)
-}
-
-export const jaroWinklerSimilarity = (a: string, b: string): number => {
-  return pkg.jaro_winkler_sim(a, b)
-}
-
-export const levenshteinDistance = (a: string, b: string): number => {
-  return pkg.levenshtein_dist(a, b)
-}
-
 type ArrayOf<T> = { push: (item: T) => void }
 const fill = <T, A extends ArrayOf<T>>(arr: A, items: T[]) => {
   items.forEach((item) => arr.push(item))
